Consolidate source-file-generator/converter into soy-generator module
